Resolve issues #814, #815, #819, #823: governance refactor, unified token adapter, math library expansion, e2e tests
- #823: Refactor governance module into composable primitives
  - Split monolithic governance.rs (2045 lines) into focused submodules:
    proposal.rs, voting.rs, simulation.rs, multisig.rs, guardians.rs,
    recovery.rs, analytics.rs, execution.rs
  - Fixed duplicated execute_proposal_type (governance.rs & timelock.rs)
    by creating shared execution.rs module
  - Fixed flash loan protection bug where snapshot-based voting power
    was computed but immediately overwritten with live balance
  - Maintained full backward compatibility via re-exports in mod.rs

- #819: Standardize token handling with unified token adapter pattern
  - Created new contracts/token crate with UnifiedTokenAdapter
  - Defines unified TokenAdapter trait for all token operations
  - Consolidates ERC-20, native, and wrapped token handling
  - Includes dust filtering, decimal normalization, batch operations
  - Added to workspace dependencies

- #815: Extract common math library for interest and liquidation
  - Added lending.rs module to contracts/math with domain-specific functions:
    InterestRateModel, health_factor_bps, is_liquidatable, ltv_bps,
    max_liquidatable, liquidation_bonus, seize_amount, validate_collateral_ratio
  - Consolidates scattered interest/liquidation calculations

- #814: Build end-to-end test suite for complete user journey
  - Added comprehensive-journey.e2e.test.ts covering:
    Multi-user concurrent operations, admin pause/resume,
    oracle price impact, liquidation with close factor enforcement,
    edge cases, multi-collateral lifecycle

Closes #814, Closes #815, Closes #819, Closes #823
